Cleanup, fixed warnings, removed makefile symlink
-Lots of removed unnecessary code
-Fixed a good chunk of the compiler warnings, see Gamecube/build.log for details
-Removed sym link, as it doesn't work on all systems and seems pointless

+++ misc.c
@@ -312,7 +312,7 @@ int CheckCdrom() {
 		if (exename[i - 2] == ';') i-= 2;
 		c = 8; i--;
 		while (i >= 0 && c >= 0) {
-			if (isalnum(exename[i])) CdromId[c--] = exename[i];
+			if (isalnum((int) exename[i])) CdromId[c--] = exename[i];
 			i--;
 		}
 	}
